backend/requirements.txt
email-validator==2.3.0
fastapi==0.135.2
joblib==1.5.2
music21==9.9.1
motor==3.7.1
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-multipart==0.0.22
scikit-learn==1.7.2
uvicorn[standard]==0.42.0
